Add WriteFloat, just an alias for WriteDouble

+++ test.cpp
@@ -65,7 +65,7 @@ void output_foo( JsonWriteScope* parent_scope, Foo* foo, const char* foo_name )
 {
   JsonWriteScopeObject json( parent_scope, foo_name );
 
-  json.WriteDouble( foo->a, "a" ); 
+  json.WriteFloat( foo->a, "a" ); 
   json.WriteInt( foo->b, "b" ); 
   json.WriteUInt( foo->b, "ub" );
